+++ examples/simple/search_indexes/documents/__init__.py
@@ -1,2 +1,3 @@
+from .author import *
 from .book import *
 from .publisher import *