+++ examples/simple/settings/base.py
@@ -270,6 +270,7 @@ ELASTICSEARCH_DSL = {
 
 # Name of the Elasticsearch index
 ELASTICSEARCH_INDEX_NAMES = {
+    'search_indexes.documents.author': 'author',
     'search_indexes.documents.book': 'book',
     'search_indexes.documents.publisher': 'publisher',
 }