+++ examples/simple/settings/testing.py
@@ -14,6 +14,7 @@ DATABASES = {
 
 # Name of the Elasticsearch index
 ELASTICSEARCH_INDEX_NAMES = {
+    'search_indexes.documents.author': 'test_author',
     'search_indexes.documents.book': 'test_book',
     'search_indexes.documents.publisher': 'test_publisher',
 }