+++ scripts/free_ports.sh
@@ -1,0 +1,2 @@
+sudo fuser -n tcp -k 9200
+sudo fuser -n tcp -k 8000